+++ src/Fp/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this package will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
@@ -7,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
+
 - `MemAnnotation` record struct for memory annotations
 - `Processor.DecodeHex(ReadOnlySpan<char>, Span<byte>, bool)` overload
 - `Processor.GetHexByteCount(ReadOnlySpan<char>, out int)`
@@ -14,22 +16,29 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Processor.LogChunk(string, bool)`
 - `string StringBuilderLog.Delimiter { get; set; }` for control over default line delimiter
 - `StringBuilderLog.GetContent()` / `StringBuilderLog.Clear()`
+- `IChunkWriter`
 
 ## Changed
+
 - `StringData` is now a record struct
 - `DataUtil` renamed to `BufferExtensions`
 - `Processor.DecodeHex(string, bool)` now uses `ReadOnlySpan<char>` instead of `string`
 - `ProcessorConfiguration` constructor parameter order adjusted with defaults for most args
-- `ProcessorConfiguration.LogReceiver` is nullable
+- `ILogReceiver ProcessorConfiguration.LogReceiver` is now `ILogWriter? ProcessorConfiguration.LogWriter`
 - `Processor` now uses `NullLog.Instance` as default log receiver
 - `Processor` now uses `Array.Empty<string>()` as default args
+- Hex writing now uses `IChunkWriter` interface type
+- `ILogReceiver.LogLevel` moved to outer scope
+- `ILogReceiver` renamed to `ILogWriter` with base interface `IChunkWriter`
 
 ## [0.26.0] 2022-1-30
 
 ### Added
+
 - FileProcessorInfo (moved from Fp.Fs FsProcessorInfo)
 
 ### Changed
+
 - Arg setup now happens in `Processor.Prepare(ProcessorConfiguration?)`
 - FormatProcessor is now derived from FileProcessor
 - Adjusted FormatProcessor members (moved commons to FileProcessor)
@@ -37,18 +46,23 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.25.0] 2022-1-18
 
 ### Added
+
 - Additional filename members in FormatProcessor
 
 ### Changed
+
 - FormatProcessor.Name now FormatProcessor.InputFile (parity with FsProcessor)
 
 ### Removed
+
 - ProcessorException
 - Scripting class
 - Filesystem-related code (moved to Fp.Fs)
 
 ## [0.24.0] 2022-1-16
+
 ### Added
+
 - WriteContext extensions in ReadWriteContextByteExtensions
 - `WriteContext<byte>.GetAdvance(int)` to claim buffer
 - Clearing of stored singletons via `One<T>.Clear` / and `One.ClearAll`
@@ -58,6 +72,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Processor.Configuration` for sharing
 
 ### Changed
+
 - Fix xor arm intrinsic
 - Initialize Processor.LittleEndian to true in constructor
 - Use local variables in static Processor.GetX/SetX
@@ -68,24 +83,33 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Moved `T FsProcessor.Initialize<T>(string[]? args)` to Processor
 
 ### Removed
+
 - `DataProcessor` / `LayeredDataProcessor` / `ProcessorChild`
 
 ## [0.23.2] 2022-1-11
+
 ### Changed
+
 - Fix intro printing on Windows
 
 ## [0.23.1] 2021-12-27
+
 ### Changed
+
 - Fix parallel check in InitializeProcessors
 
 ## [0.23.0] 2021-12-27
+
 ### Changed
+
 - Update xmldoc
 - Parallel value 0 (default) corresponds to synchronous execution, otherwise interpret as async of n workers
 - Always dispose stream for parallel-mode FSS read
 
 ## [0.22.0] 2021-12-17
+
 ### Changed
+
 - Miscellaneous API changes
 - Internal fixes
 - Changed RealFileSystemSource to public
@@ -95,11 +119,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed incorrect documentation on AlignUp/AlignDown
 
 ## [0.21.1] 2021-11-9
+
 ### Changed
+
 - Fixed console output on Windows
 
 ## [0.21.0] 2021-11-9
+
 ### Changed
+
 - Updated / fixed Coordinator API
 - Add length-based precedence to `ValidateExtension`
 - Respect endianness on unmanaged helpers for `ReadOnlySpan<byte>`